+++ hibike/devices/LimitSwitch/limit_switch.cpp
@@ -40,7 +40,7 @@ uint32_t device_write(uint8_t param, uint8_t* data, size_t len){
 //
 //    return          -   sizeof(param) on success; 0 otherwise
 
-uint8_t device_data_update(int param, uint8_t* data_update_buf, size_t buf_len) {
+uint8_t device_data_update(uint8_t param, uint8_t* data_update_buf, size_t buf_len) {
 
   if (MAX_PAYLOAD_SIZE - buf_len < sizeof(uint8_t) || param >= NUM_SWITCHES) {
     return 0;
