+++ hibike/devices/YogiBear/YogiBear.cpp
@@ -201,7 +201,7 @@ uint32_t device_write(uint8_t param, uint8_t* data, size_t len){
 // append_buf copies the specified amount data into the dst buffer and increments the offset
 
 
-uint8_t device_data_update(int param, uint8_t* data_update_buf, size_t buf_len) {
+uint8_t device_data_update(uint8_t param, uint8_t* data_update_buf, size_t buf_len) {
 
   switch (param) {
     case PARAM_DUTY:
